EditView: Rename get_save_button_sensitivity() to set_save_button_sensitivity()
The name 'get_save_button_sensitivity' sounds like as if we're
returning the actual value of `save_button.sensitive`.
However, what we essensially do here is decide if we should
make `save_button` sensitive or not and set that returned value
to `save_button.sensitive`. So maybe the name `set_*` instead of
`get_*` should make more sense here.